Dashboard/requirements.txt
flask==2.0.1
finnhub-python==2.4.13
python-dotenv==0.19.0
pandas==1.3.3